+++ src/raw/io/output/into_reader/mod.rs
@@ -39,3 +39,6 @@ pub use by_range::by_range;
 
 mod by_offset;
 pub use by_offset::by_offset;
+
+mod from_fd;
+pub use from_fd::from_fd;
